requirements.txt
numpy~=2.2.6
opencv-python~=4.13.0.92
scikit-image~=0.25.2
scipy~=1.15.3
matplotlib~=3.10.8
pandas~=2.3.3
pillow~=12.1.1
streamlit~=1.55.0
streamlit-drawable-canvas-fix==0.9.8
